Transferred this file from backend, cleaned it up and got it working
